Add ability to register res4d timeseries
This can be useful for removing confound information
(e.g. for functional connectivity analysis). The current version
is unfinished. The timeseries is not getting renamed properly
in the data sink, and I also have not determined whether this
is the right implementation. (Using a command-line switch means
there's no record of whether the timeseries that ends up in the
reg/ directory is pre- or post- model fit, but it doesn't make
sense to parameterize this at the experiment level.