Fix to FSL highpass filter changes
This works around some changes in later version of FSL (as of 11/2014)
that return a de-meaned timeseries from the FSL highpass filter.
In FEAT, the mean is replaced, and the rest of the processing carries
on as usual. Because I don't want to break compatability with older
versions of FSL, this adds back in the mean but only if it looks
like the filtered timeseries has been de-meaned.