Add cross-experiment epi space registration
This will be useful for projects that use functional localizers and similar.
The registration is accomplished through the ``-regexp`` flag in
``run_fmri.py``. For example, the cmdline

```
run_fmri.py -exp A -regexp B -regspace epi -timeseries
```

will combine the func-to-anat matrices from experiment A and the
anat-to-func matrix from the first run of experiment B, placing the
experiment A files in a common space with experiment B files.