Split off slice time correction into sub workflow; add siemens interleaved
Siemens slice order when the acquisition is interleaved depends on the
number of slices. This change uses the custom slice order option to handle
that weirdness (but does not otherwise expose custom slice order/timing,
so it probably still won't work for SMS sequences).

See this blog post for more information:
https://practicalfmri.blogspot.com/2012/07/siemens-slice-ordering.html